refactor(components): refactor image (#3770)
